Extract valueEqual() to its own file; tighten its param types
Move `valueEqual()` out of `fabric-value.ts` into a dedicated
`valueEqual.ts`, re-exported through the `fabric-value` barrel. This
leaves `fabric-value.ts` a pure re-export surface and gives the single
function a file whose name matches the export.

Tighten the parameters from `unknown` to `FabricValue`, so the public
type states the domain the function actually operates over. The two
callers in `scheduler/diagnosis.ts` hold stored read/write values typed
`unknown` in the surrounding storage plumbing; they cast to
`FabricValue` at the call, which is sound (these are stored fabric
values) and behavior-preserving.
